moves twitter embed to own component
